pyproject.toml
[project]
name = "torch-stuff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torchvision>=0.21.0",
]
